+++ docs/lib/search/search-types.ts
